Remove duplicated code and dead code, and moved topk functions back to topkgating for better clarity
